Print an unsigned rank as one, on every platform
pmix_util_print_rank() names five reserved ranks and renders everything
else numerically. PMIX_RANK_VALID is UINT32_MAX-50, so the reserved band
holds fifty values and forty-five of them reach that numeric branch -
which cast the uint32_t to long and printed it with %ld. That is exact
where long is 64 bits and implementation-defined where it is 32, so the
same rank appeared in one platform's diagnostics as a large positive
number and in another's as a negative one. Render it with %u, which is
exact for a uint32_t everywhere.

pmix_name_fns.h also guarded its <stdint.h> include on HAVE_STDINT_h
rather than HAVE_STDINT_H. The guard was therefore always false and the
include never happened; the header builds only because pmix_common.h,
two lines below, brings the type in anyway. That is precisely the silent
misspelling the project's "test with #if, not #ifdef" rule exists to
prevent, and it should not be left waiting for a platform where the
transitive include is absent.

While here, write down two contracts the header had never stated. The
print helpers answer out of a 16-entry per-thread ring, but a name print
spends two entries of it - one for the rank, one for the assembled pair
- so the budget for a single statement is about eight; and they answer
with a shared static string rather than NULL when they cannot get a
buffer, so the result is safe to print and not safe to write through.
The comparator's three-way rank compare gets the same treatment, since
the reason it is not a subtraction is not obvious from the code.

The regression case for the numeric branch cannot fail on a 64-bit
build; it is there for the 32-bit ones, and says so.

+++ src/util/pmix_name_fns.c
@@ -214,7 +214,15 @@ char *pmix_util_print_rank(const pmix_rank_t vpid)
     } else if (PMIX_RANK_INVALID == vpid) {
         pmix_snprintf(ptr->buffers[index], PMIX_PRINT_NAME_ARGS_MAX_SIZE, "INVALID");
     } else {
-        pmix_snprintf(ptr->buffers[index], PMIX_PRINT_NAME_ARGS_MAX_SIZE, "%ld", (long) vpid);
+        /* pmix_rank_t is uint32_t and the reserved band runs from
+         * PMIX_RANK_VALID (UINT32_MAX-50) up, only five of which are
+         * named above - so a value this branch has to render can exceed
+         * LONG_MAX where long is 32 bits, and casting it there is
+         * implementation-defined: the same rank printed as a large
+         * positive number on one platform and a negative one on another.
+         * Render it as what it is. */
+        pmix_snprintf(ptr->buffers[index], PMIX_PRINT_NAME_ARGS_MAX_SIZE, "%u",
+                      (unsigned int) vpid);
     }
     ptr->cntr++;
     if (PMIX_PRINT_NAME_ARG_NUM_BUFS == ptr->cntr) {
